+++ Dockerfile
@@ -1,5 +1,5 @@
 # Stage 0, "build-stage" to build and compile the frontend
-FROM node:11-alpine as build-stage
+FROM node:11-alpine AS build-stage
 WORKDIR /app
 
 RUN npm install
@@ -7,6 +7,6 @@ COPY . ./
 RUN npm run build
 
 # Stage 1, to have only the compiled app, ready for production with Nginx
-from nginx:1.15-alpine
+FROM nginx:1.15-alpine
 COPY --from=build-stage /app/build /usr/share/nginx/html
 COPY ./nginx.conf /etc/nginx/conf.d/default.conf
