+++ src/App.tsx
@@ -4,15 +4,17 @@ import { BrowserRouter, Switch, Route, Redirect } from 'react-router-dom';
 import { Send } from 'containers/Send';
 import { Receive } from 'containers/Receive';
 
+import { homepage } from 'config';
+
 const App = () => {
 	return (
 		<BrowserRouter>
 			<Switch>
-				<Route exact path="/Information_Security_JS">
-					<Redirect to="/Information_Security_JS/send" />
+				<Route exact path={homepage}>
+					<Redirect to={homepage + '/send'} />
 				</Route>
-				<Route path="/Information_Security_JS/send" component={Send} />
-				<Route path="/Information_Security_JS/receive" component={Receive} />
+				<Route path={homepage + '/send'} component={Send} />
+				<Route path={homepage + '/receive'} component={Receive} />
 			</Switch>
 		</BrowserRouter>
 	);
