+++ src/methods/encryption.ts
